requirements.txt
nextcord
python-dotenv
opencv-python
torch